+++ installer/install-domain.js
@@ -1,7 +1,7 @@
-import { createNexusRealtimeKitInstaller } from "./kit-installer.js";
+import { createNexusEngineKitInstaller } from "./kit-installer.js";
 
 export async function installDomainKits(engine, domainId, config = {}, options = {}) {
-  const kitInstaller = createNexusRealtimeKitInstaller(options);
+  const kitInstaller = createNexusEngineKitInstaller(options);
   return kitInstaller.installDomain(engine, domainId, config);
 }