+++ installer/install-kit.js
@@ -1,5 +1,5 @@
-import { createNexusRealtimeKitInstaller } from "./kit-installer.js";
+import { createNexusEngineKitInstaller } from "./kit-installer.js";
 
 export async function installKit(engine, kitOrId, config = {}, options = {}) {
-  return createNexusRealtimeKitInstaller(options).installKit(engine, kitOrId, config);
+  return createNexusEngineKitInstaller(options).installKit(engine, kitOrId, config);
 }